requirements.txt
Flask==2.3.3
Werkzeug==2.3.6
Flask-Cors==3.1.0
Pillow==10.0.1
Tesseract==0.3.2
deep-translator==1.10.1

